requirements.txt
numpy==1.22.3
click
Flask
tensorflow
keras
scipy==1.8.0
Pillow
